terminado seccion de buscador panel administrativo y mostrar

+++ database/Creacion_Esquema_BaseDatos20211008.sql
@@ -3,8 +3,10 @@ CREATE TABLE IF NOT EXISTS `Paises` (
   `Pais` varchar(100) NOT NULL,
   PRIMARY KEY (`Codigo`)
 ) Engine = InnoDB;
+
 ALTER TABLE
   Paises CHANGE COLUMN Codigo Id varchar(2) NOT NULL;
+
 CREATE TABLE IF NOT EXISTS `Ciudades` (
     `idCiudades` int(11) NOT NULL AUTO_INCREMENT,
     `Paises_Codigo` varchar(2) NOT NULL,
@@ -13,10 +15,13 @@ CREATE TABLE IF NOT EXISTS `Ciudades` (
     KEY `Paises_Codigo` (`Paises_Codigo`),
     KEY `Ciudad` (`Ciudad`)
   ) Engine = InnoDB;
+
 ALTER TABLE
   ciudades CHANGE COLUMN idCiudades Id int(11) NOT NULL;
+
 ALTER TABLE
   ciudades CHANGE COLUMN Paises_Codigo Id_Pais varchar(2) NOT NULL;
+
 create table Usuarios(
     Id int auto_increment,
     Usuario varchar(55),
@@ -37,6 +42,7 @@ create table Usuarios(
     constraint UQ_USUARIOS_EMAIL UNIQUE (Email)
   ) Engine = InnoDB;
 use pablogarciajc_ecommerce;
+
 CREATE TABLE categorias(
     id int(255) auto_increment not null,
     categorias varchar(100) not null,
@@ -45,15 +51,14 @@ CREATE TABLE categorias(
 use pablogarciajc_ecommerce;
 CREATE TABLE productos(
     id int(255) auto_increment not null,
-    categoria_id int(255),   
+    categoria_id int(255),
     nombre varchar(100),
     descripcion text,
-    precio float(100, 2),    
+    precio float(100, 0),
     stock int(255),
-    oferta float(100,0),
+    oferta float(100, 0),
     marca varchar(50),
-    /* memoria_ram varchar(50), */
-    memoria_ram float(100,0),
+    memoria_ram float(100, 0), 
     imagen varchar(255),
     CONSTRAINT pk_categorias PRIMARY KEY(id),
     CONSTRAINT fk_producto_categoria FOREIGN KEY(categoria_id) REFERENCES categorias(id)
